requirements.txt
pytz==2025.2
requests==2.31.0
websockets==12.0
pandas==2.0.3
numpy==1.26.4
python-dotenv==1.0.0
scikit-learn==1.2.2
ta==0.10.2
setuptools==72.1.0
nltk==3.8.1
plotly==5.18.0
pytest==7.4.0
logtail-python==0.2.9
fastapi==0.111.0
uvicorn==0.30.1
slowapi==0.1.9
cryptography==43.0.1
torch==2.4.1
streamlit==1.39.0
redis==5.1.1
optuna==4.0.0
prometheus_fastapi_instrumentator==7.0.0
psycopg2-binary==2.9.10
ccxt==4.0.0
feedparser==6.0.10
pypfopt==1.5.5